Fix bug where IndentationError was causing program to stop. Fixes GH-11
We use tokenize to parse python tokens, but since we can start the
parser on a line in the middle of the file, it doesn't have indentation
context from previous lines and can throw an IndentationError. In these
cases we can safely catch and stop iteration since we haven't found
the tokens we are looking for anyway.